+++ changelog.md
@@ -1,1 +1,7 @@
 * Update depends.
+* Remove modify checkleg command.
+* The config directory has been changed and performance has been improved.
+* Update storage.
+* Change command checkleg to checklegendary and checkleg (aliases).
+* HEX color support has been added.
+* Other minor improvements and changes.